// web/src/app/api/generated/src/models/SpecialistAppointmentExecutionContext.ts
export interface SpecialistAppointmentExecutionContext {
  /**
   *
   * @type {string}
   * @memberof SpecialistAppointmentExecutionContext
   */
  appointmentId?: string;
  /**
   *
   * @type {string}
   * @memberof SpecialistAppointmentExecutionContext
   */
  appointmentStatus?: string;
  /**
   *
   * @type {string}
   * @memberof SpecialistAppointmentExecutionContext
   */
  participantId?: string;
  /**
   *
   * @type {string}
   * @memberof SpecialistAppointmentExecutionContext
   */
  participantName?: string;
  /**
   *
   * @type {string}
   * @memberof SpecialistAppointmentExecutionContext
   */
  planId?: string;
  /**
   *
   * @type {string}
   * @memberof SpecialistAppointmentExecutionContext
   */
  plannedSessionId?: string;
  /**
   *
   * @type {Array<SpecialistExecutionPrescriptionView>}
   * @memberof SpecialistAppointmentExecutionContext
   */
  prescriptions?: Array<SpecialistExecutionPrescriptionView>;
  /**
   *
   * @type {SpecialistExecutionSummary}
   * @memberof SpecialistAppointmentExecutionContext
   */
  recordedExecution?: SpecialistExecutionSummary;
  /**
   *
   * @type {boolean}
   * @memberof SpecialistAppointmentExecutionContext
   */
  recordingAllowed?: boolean;
  /**
   *
   * @type {string}
   * @memberof SpecialistAppointmentExecutionContext
   */
  revisionId?: string;
  /**
   *
   * @type {string}
   * @memberof SpecialistAppointmentExecutionContext
   */
  sessionTitle?: string;
}