// web/src/app/api/generated/src/models/SpecialistExecutionPrescriptionView.ts
export interface SpecialistExecutionPrescriptionView {
  /**
   *
   * @type {string}
   * @memberof SpecialistExecutionPrescriptionView
   */
  canonicalDoseType?: string;
  /**
   *
   * @type {number}
   * @memberof SpecialistExecutionPrescriptionView
   */
  contacts?: number;
  /**
   *
   * @type {number}
   * @memberof SpecialistExecutionPrescriptionView
   */
  distanceMeters?: number;
  /**
   *
   * @type {string}
   * @memberof SpecialistExecutionPrescriptionView
   */
  doseType?: string;
  /**
   *
   * @type {number}
   * @memberof SpecialistExecutionPrescriptionView
   */
  durationSeconds?: number;
  /**
   *
   * @type {string}
   * @memberof SpecialistExecutionPrescriptionView
   */
  exerciseName?: string;
  /**
   *
   * @type {string}
   * @memberof SpecialistExecutionPrescriptionView
   */
  externalLoadUnit?: string;
  /**
   *
   * @type {number}
   * @memberof SpecialistExecutionPrescriptionView
   */
  externalLoadValue?: number;
  /**
   *
   * @type {string}
   * @memberof SpecialistExecutionPrescriptionView
   */
  id?: string;
  /**
   *
   * @type {string}
   * @memberof SpecialistExecutionPrescriptionView
   */
  intensityType?: string;
  /**
   *
   * @type {number}
   * @memberof SpecialistExecutionPrescriptionView
   */
  intensityValue?: number;
  /**
   *
   * @type {string}
   * @memberof SpecialistExecutionPrescriptionView
   */
  intensityZone?: string;
  /**
   *
   * @type {string}
   * @memberof SpecialistExecutionPrescriptionView
   */
  notes?: string;
  /**
   *
   * @type {number}
   * @memberof SpecialistExecutionPrescriptionView
   */
  position?: number;
  /**
   *
   * @type {number}
   * @memberof SpecialistExecutionPrescriptionView
   */
  repetitions?: number;
  /**
   *
   * @type {number}
   * @memberof SpecialistExecutionPrescriptionView
   */
  restSeconds?: number;
  /**
   *
   * @type {number}
   * @memberof SpecialistExecutionPrescriptionView
   */
  sets?: number;
  /**
   *
   * @type {string}
   * @memberof SpecialistExecutionPrescriptionView
   */
  side?: string;
  /**
   *
   * @type {string}
   * @memberof SpecialistExecutionPrescriptionView
   */
  tempo?: string;
}